src/data-repository/models/user.ts
export interface User {
  _id: number;
  url: string;
  external_id: string;
  name: string;
  alias?: string;
  created_at: string;
  active: boolean;
  verified?: boolean;
  shared: boolean;
  locale?: string;
  timezone?: string;
  last_login_at: string;
  email?: string;
  phone: string;
  signature: string;
  organization_id?: number;
  tags: string[];
  suspended: boolean;
  role: string;
}

export const getUser = (props: User) => ({
  ...props,
  getOrganization: () => {},
  getTickets: () => {},
  getSearchableFields: () => Object.keys(props),
});
